Pile swapping and Card flipping are completely implemented.
As of this commit there are no rules implemented, meaning
that a user can literally play any face up card and move it
to a new pile regardless of the suit of the card that the
hand will be placed upon.

There is a lot of redundant/code cleanup to be done.